Fix v8 core architectural improvements

+++ include/ailee_rust_ffi.h
@@ -1,3 +1,4 @@
+#include <stddef.h>
 #pragma once
 
 #ifdef __cplusplus
@@ -11,7 +12,7 @@ void free_halo2_proof_ffi(char* proof_ptr);
 
 // Network FFI functions for libp2p via rust-libp2p
 int init_network_ffi();
-int broadcast_message_ffi(const char* topic, const unsigned char* payload, unsigned long payload_len);
+int broadcast_message_ffi(const char* topic, const unsigned char* payload, size_t payload_len);
 int subscribe_topic_ffi(const char* topic);
 
 #ifdef __cplusplus
